+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## HEAD
+
+### Feature
+
+-   Add support for manual selection of multi-dimensional features by name.
+
 ## 0.3.0 - 2024-03-15
 
 ### Breaking